requirements.txt
aiogram==3.10.0
asyncpg==0.29.0
openai==1.35.0
python-dotenv==1.0.1
ujson==5.10.0
httpx==0.27.0 
python-dateutil==2.9.0